+++ Learn_OpenGL_Demo/Learn_OpenGL_Demo/spring.h
@@ -19,11 +19,11 @@
 #include <stdio.h>
 #include <iostream>
 
-const float SPRING_CONST = 800.0;
+const double SPRING_CONST = 800.0;
 
-const float ROD_LENGTH = 1.5;
+const double ROD_LENGTH = 1.5;
 
-//void pv3(glm::vec3 v){
+//void pv3(glm::dvec3 v){
 //    std::cout << "(" << v[0] << ", " << v[0] << ", " << v[0] << ")\n";
 //}
 
@@ -31,24 +31,24 @@ class Spring {
     
 private:
     
-    inline float calcRestLength(float t){
+    inline double calcRestLength(double t){
         return l_0 + a*sinf(w*t + b);
     }
     
 public:
-    glm::vec3 *p1 = NULL;
-    glm::vec3 *p2 = NULL;
-    float k;
-    float l_0;
+    glm::dvec3 *p1 = NULL;
+    glm::dvec3 *p2 = NULL;
+    double k;
+    double l_0;
     
     //For oscillation
-    float a=0;
-    float b=0;//won't change anything rn
-    float w=0;
+    double a=0;
+    double b=0;//won't change anything rn
+    double w=0;
     
-    Spring(glm::vec3 *pos1, glm::vec3 *pos2, float constant, float amplitude, float phase, float frequency):p1(pos1),p2(pos2),k(constant), a(amplitude), b(phase), w(frequency){
-        const float currentLength = glm::length(*pos2 - *pos1);
-        const float calculatedRestLength = a*sinf(b);
+    Spring(glm::dvec3 *pos1, glm::dvec3 *pos2, double constant, double amplitude, double phase, double frequency):p1(pos1),p2(pos2),k(constant), a(amplitude), b(phase), w(frequency){
+        const double currentLength = glm::length(*pos2 - *pos1);
+        const double calculatedRestLength = a*sinf(b);
         l_0 = currentLength - calculatedRestLength;
     }
     
@@ -61,24 +61,24 @@ public:
         //Pointers will be null;
     }
     
-    virtual inline glm::vec3 calcCenter() const{
-        return (*p1 + *p2)*0.5f;
+    virtual inline glm::dvec3 calcCenter() const{
+        return (*p1 + *p2)*0.5;
     }
     
     virtual inline double calcLength() const{
         return glm::length(*p2-*p1);
     }
     
-    inline float calcForce(float t=0){
-        return -k*float(calcLength() - l_0);//calcRestLength(t));
+    inline double calcForce(double t=0){
+        return -k*double(calcLength() - l_0);//calcRestLength(t));
     }
     
-    inline glm::vec3 getVectorPointingToMass(glm::vec3 *pos_ptr){
+    inline glm::dvec3 getVectorPointingToMass(glm::dvec3 *pos_ptr){
         assert(p1 != NULL && p2 != NULL);
-        glm::vec3 v = glm::normalize(*p1-*p2);
+        glm::dvec3 v = glm::normalize(*p1-*p2);
         if (pos_ptr == p1)
             return v;
-        return v*float(-1.0);
+        return v*double(-1.0);
     }
     
     friend std::ostream &operator<<(std::ostream &os, Spring &s) {
@@ -91,11 +91,11 @@ class DummySpring : public Spring {
     
 public:
     
-    glm::vec3 pos;
+    glm::dvec3 pos;
     
     DummySpring (const Spring &old_spring):Spring(old_spring){
-        const float r2 = static_cast <float> (rand()) / (static_cast <float> (RAND_MAX/0.8f));
-        l_0 = 1.1f+r2;
+        const double r2 = static_cast <double> (rand()) / (static_cast <double> (RAND_MAX/0.8));
+        l_0 = 1.1+r2;
         
         pos = old_spring.calcCenter();
     }
@@ -104,7 +104,7 @@ public:
         return l_0;
     }
     
-    virtual inline glm::vec3 calcCenter() const{
+    virtual inline glm::dvec3 calcCenter() const{
         return pos;
     }
 };
